refactor: 상수 데이터 파일로 분리
