+++ packages/client/src/constants/index.ts
@@ -1,1 +1,2 @@
 export * from './colors';
+export { JOB_DICT } from './job';
